requirements.txt
openai
pyTelegramBotAPI
tiktoken
Pillow